requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.65.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
